// frontend/src/api/generated/models/if-execution-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { IfExecutionStatus } from './if-execution-status';
// May contain unused imports in some cases
// @ts-ignore
import type { ImportMode } from './import-mode';

/**
 * 
 * @export
 * @interface IfExecutionItem
 */
export interface IfExecutionItem {
    /**
     * 実行履歴ID
     * @type {number}
     * @memberof IfExecutionItem
     */
    'id': number;
    /**
     * I/F種別（IFX-001 / IFX-002）
     * @type {string}
     * @memberof IfExecutionItem
     */
    'ifType': string;
    /**
     * 元ファイル名
     * @type {string}
     * @memberof IfExecutionItem
     */
    'fileName': string;
    /**
     * CSVデータ総行数
     * @type {number}
     * @memberof IfExecutionItem
     */
    'totalCount': number;
    /**
     * バリデーション成功行数
     * @type {number}
     * @memberof IfExecutionItem
     */
    'successCount': number;
    /**
     * バリデーションエラー行数
     * @type {number}
     * @memberof IfExecutionItem
     */
    'errorCount': number;
    /**
     * 
     * @type {ImportMode}
     * @memberof IfExecutionItem
     */
    'mode': ImportMode;
    /**
     * 
     * @type {IfExecutionStatus}
     * @memberof IfExecutionItem
     */
    'status': IfExecutionStatus;
    /**
     * Blob移動失敗フラグ
     * @type {boolean}
     * @memberof IfExecutionItem
     */
    'blobMoveFailed': boolean;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof IfExecutionItem
     */
    'warehouseId': number;
    /**
     * 実行日時
     * @type {string}
     * @memberof IfExecutionItem
     */
    'executedAt': string;
    /**
     * 実行ユーザーID
     * @type {number}
     * @memberof IfExecutionItem
     */
    'executedBy': number;
    /**
     * 実行ユーザー氏名
     * @type {string}
     * @memberof IfExecutionItem
     */
    'executedByName': string;
}