// frontend/src/api/generated/models/if-execution-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { IfExecutionItem } from './if-execution-item';

/**
 * 
 * @export
 * @interface IfExecutionPageResponse
 */
export interface IfExecutionPageResponse {
    /**
     * 
     * @type {Array<IfExecutionItem>}
     * @memberof IfExecutionPageResponse
     */
    'content': Array<IfExecutionItem>;
    /**
     * 
     * @type {number}
     * @memberof IfExecutionPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof IfExecutionPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof IfExecutionPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof IfExecutionPageResponse
     */
    'totalPages': number;
}